Refactor component types
