+++ src/components/Counter/index.tsx
@@ -1,12 +1,6 @@
 import './Counter.css';
 
-export function Counter({
-	total,
-	completed,
-}: {
-	total: number;
-	completed: boolean;
-}) {
+export function Counter({ total, completed }: Counter) {
 	return (
 		<h4>
 			You&apos;ve completed {completed} of {total} TODOs
