+++ src/components/ListItem/index.tsx
@@ -1,12 +1,5 @@
 import './ListItem.css';
 
-interface ListItem {
-	text: string;
-	done: boolean;
-	toggleComplete: () => void;
-	removeTodo: () => void;
-}
-
 export function ListItem({ text, done, toggleComplete, removeTodo }: ListItem) {
 	return (
 		<li className="TodoItem">
